requirements.txt
garminconnect
python-dotenv
google-genai
requests
gspread
oauth2client